# Conflicts:
#	BiBot/card-service/src/main/java/com/coderecipe/v1/payment/dto/vo/PaymentReq.java
#	BiBot/card-service/src/main/java/com/coderecipe/v1/payment/service/IPaymentHistoryService.java
#	BiBot/card-service/src/main/java/com/coderecipe/v1/payment/service/impl/PaymentHistoryImpl.java

+++ BiBot/card-service/src/main/java/com/coderecipe/v1/card/service/impl/CardServiceImpl.java
@@ -9,8 +9,10 @@ import com.coderecipe.v1.card.model.Card;
 import com.coderecipe.v1.card.model.repository.ICardRepository;
 import com.coderecipe.v1.card.service.ICardService;
 import com.coderecipe.v1.payment.dto.vo.PaymentRes.*;
+
 import com.coderecipe.v1.payment.model.repository.IPaymentHistoryRepository;
 
+import java.time.LocalDateTime;
 import java.util.UUID;
 
 import lombok.Data;
@@ -30,8 +32,8 @@ public class CardServiceImpl implements ICardService {
     private final IPaymentHistoryRepository iPaymentHistoryRepository;
 
     @Override
-    public Long addCard(CreateCard req) {
-        Card card = Card.of(req);
+    public Long addCard(CreateCard req, UUID userId) {
+        Card card = Card.of(req, userId);
         iCardRepository.save(card);
         return card.getId();
     }
@@ -39,25 +41,41 @@ public class CardServiceImpl implements ICardService {
     @Override
     public CardDTO getCard(Long cardId) {
         Card card = iCardRepository.findById(cardId)
-                .orElseThrow(() -> new CustomException(ResCode.CARD_NOT_FOUND));
+            .orElseThrow(() -> new CustomException(ResCode.CARD_NOT_FOUND));
         return CardDTO.of(card);
     }
 
     @Override
     public List<CardInfoRes> getAllCard(UUID userId) {
         return iCardRepository.findAllByUserIdOrderById(userId)
-                .stream().map(CardInfoRes::of).toList();
+            .stream().map(CardInfoRes::of).toList();
     }
 
     @Override
-    public Long deleteCard(Long cardId) {
-        iCardRepository.deleteById(cardId);
-        return cardId;
+    public Long deleteCard(Long cardId, UUID userId) {
+        Card card = iCardRepository.findById(cardId)
+                .orElseThrow(() -> new CustomException(ResCode.CARD_NOT_FOUND));
+        if (card.getUserId() != userId) {
+            throw new CustomException(ResCode.BAD_REQUEST);
+        } else {
+            iCardRepository.deleteById(cardId);
+            return cardId;
+        }
     }
 
     @Override
-    public List<PaymentInfo> getPayments(Long cardId) {
-        return iPaymentHistoryRepository.findAllByCardId(cardId)
-                .stream().map(PaymentInfo::of).toList();
+    public List<PaymentInfo> getPayments(Long cardId, LocalDateTime startDateTime, LocalDateTime endDateTime) {
+        return iPaymentHistoryRepository.findAllByRegTimeBetweenAndCardIdOrderByRegTimeDesc(
+                startDateTime, endDateTime, cardId).stream()
+            .map(PaymentInfo::of).toList();
     }
+
+
+    @Override
+    public Integer getAmount(Long cardId, LocalDateTime startDateTime, LocalDateTime endDateTime) {
+        return iPaymentHistoryRepository.findAndSumAllByRegTimeBetweenAndCardId(
+            startDateTime, endDateTime,
+            iCardRepository.getReferenceById(cardId));
+    }
+
 }
